Antenna toolkit improve architecture
