+++ pyaedt/toolkits/antennas/__init__.py
@@ -1,12 +1,12 @@
 from pyaedt.generic.general_methods import generate_unique_name
-from pyaedt.toolkits.antennas.patch import Patch
+from pyaedt.toolkits.antennas.patch import RectangularPatchProbe
 
 
 class Antennas(object):
     def __init__(self, app):
         self._app = app
 
-    def rectangular_patch_w_probe(
+    def rectangular_patch_probe(
         self,
         frequency=10.0,
         frequency_unit="GHz",
@@ -22,10 +22,9 @@ class Antennas(object):
         if not position:
             position = [0, 0, 0]
 
-        if not antenna_name:
-            antenna_name = generate_unique_name("Patch")
+        antenna_name = self._check_antenna_name(antenna_name)
 
-        rect_patch = Patch(
+        rect_patch = RectangularPatchProbe(
             self._app,
             frequency=frequency,
             frequency_unit=frequency_unit,
@@ -38,5 +37,13 @@ class Antennas(object):
             antenna_name=antenna_name,
             position=position,
         )
-        rect_patch._draw()
+        rect_patch.draw()
         return rect_patch
+
+    def _check_antenna_name(self, antenna_name=None):
+        """Check if antenna name is repeated or assign a random antenna name."""
+        if not antenna_name or self._app.modeler.oeditor.GetObjectsInGroup(antenna_name).count > 0:
+            antenna_name = generate_unique_name("Patch")
+            while self._app.modeler.oeditor.GetObjectsInGroup(antenna_name).count > 0:
+                antenna_name = generate_unique_name("Patch")
+        return antenna_name
